+++ utils/time.go
@@ -1,0 +1,5 @@
+package utils
+
+const (
+	TimeFormat = "2006/01/02 15:04:05"
+)